+++ cli/build.rs
@@ -54,6 +54,10 @@ fn create_compiler_snapshot(
   custom_libs
     .insert("lib.deno.web.d.ts".to_string(), deno_web::get_declaration());
   custom_libs.insert(
+    "lib.deno.fetch.d.ts".to_string(),
+    deno_fetch::get_declaration(),
+  );
+  custom_libs.insert(
     "lib.deno.window.d.ts".to_string(),
     cwd.join("dts/lib.deno.window.d.ts"),
   );
@@ -112,6 +116,10 @@ fn main() {
   println!(
     "cargo:rustc-env=DENO_WEB_LIB_PATH={}",
     deno_web::get_declaration().display()
+  );
+  println!(
+    "cargo:rustc-env=DENO_FETCH_LIB_PATH={}",
+    deno_fetch::get_declaration().display()
   );
 
   println!(
