+++ cli/tsc/99_main_compiler.js
@@ -435,6 +435,7 @@ delete Object.prototype.__proto__;
   ts.libs.push("deno.ns", "deno.window", "deno.worker", "deno.shared_globals");
   ts.libMap.set("deno.ns", "lib.deno.ns.d.ts");
   ts.libMap.set("deno.web", "lib.deno.web.d.ts");
+  ts.libMap.set("deno.fetch", "lib.deno.fetch.d.ts");
   ts.libMap.set("deno.window", "lib.deno.window.d.ts");
   ts.libMap.set("deno.worker", "lib.deno.worker.d.ts");
   ts.libMap.set("deno.shared_globals", "lib.deno.shared_globals.d.ts");
@@ -448,6 +449,10 @@ delete Object.prototype.__proto__;
   );
   SNAPSHOT_HOST.getSourceFile(
     `${ASSETS}/lib.deno.web.d.ts`,
+    ts.ScriptTarget.ESNext,
+  );
+  SNAPSHOT_HOST.getSourceFile(
+    `${ASSETS}/lib.deno.fetch.d.ts`,
     ts.ScriptTarget.ESNext,
   );
   SNAPSHOT_HOST.getSourceFile(
